// frontend/web/src/services/API/proserve-wb-publishing-api/models/GetAmisResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Publishing API
 * API in charge of publishing workbench products
 *
 * The version of the OpenAPI document: 2023-01-26T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Ami } from './Ami';
import {
    AmiFromJSON,
    AmiFromJSONTyped,
    AmiToJSON,
} from './Ami';

/**
 * 
 * @export
 * @interface GetAmisResponse
 */
export interface GetAmisResponse {
    /**
     * 
     * @type {Array<Ami>}
     * @memberof GetAmisResponse
     */
    amis?: Array<Ami>;
}

/**
 * Check if a given object implements the GetAmisResponse interface.
 */
export function instanceOfGetAmisResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetAmisResponseFromJSON(json: any): GetAmisResponse {
    return GetAmisResponseFromJSONTyped(json, false);
}

export function GetAmisResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetAmisResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'amis': !exists(json, 'amis') ? undefined : ((json['amis'] as Array<any>).map(AmiFromJSON)),
    };
}

export function GetAmisResponseToJSON(value?: GetAmisResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'amis': value.amis === undefined ? undefined : ((value.amis as Array<any>).map(AmiToJSON)),
    };
}