// frontend/web/src/services/API/proserve-wb-publishing-api/models/GetLatestMajorVersionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Publishing API
 * API in charge of publishing workbench products
 *
 * The version of the OpenAPI document: 2023-01-26T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { VersionSummary } from './VersionSummary';
import {
    VersionSummaryFromJSON,
    VersionSummaryFromJSONTyped,
    VersionSummaryToJSON,
} from './VersionSummary';

/**
 * 
 * @export
 * @interface GetLatestMajorVersionsResponse
 */
export interface GetLatestMajorVersionsResponse {
    /**
     * 
     * @type {Array<VersionSummary>}
     * @memberof GetLatestMajorVersionsResponse
     */
    versions: Array<VersionSummary>;
}

/**
 * Check if a given object implements the GetLatestMajorVersionsResponse interface.
 */
export function instanceOfGetLatestMajorVersionsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "versions" in value;

    return isInstance;
}

export function GetLatestMajorVersionsResponseFromJSON(json: any): GetLatestMajorVersionsResponse {
    return GetLatestMajorVersionsResponseFromJSONTyped(json, false);
}

export function GetLatestMajorVersionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetLatestMajorVersionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'versions': ((json['versions'] as Array<any>).map(VersionSummaryFromJSON)),
    };
}

export function GetLatestMajorVersionsResponseToJSON(value?: GetLatestMajorVersionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'versions': ((value.versions as Array<any>).map(VersionSummaryToJSON)),
    };
}