// frontend/web/src/services/API/proserve-wb-publishing-api/models/GetLatestTemplateResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Publishing API
 * API in charge of publishing workbench products
 *
 * The version of the OpenAPI document: 2023-01-26T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetLatestTemplateResponse
 */
export interface GetLatestTemplateResponse {
    /**
     * Template content
     * @type {string}
     * @memberof GetLatestTemplateResponse
     */
    template: string;
}

/**
 * Check if a given object implements the GetLatestTemplateResponse interface.
 */
export function instanceOfGetLatestTemplateResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "template" in value;

    return isInstance;
}

export function GetLatestTemplateResponseFromJSON(json: any): GetLatestTemplateResponse {
    return GetLatestTemplateResponseFromJSONTyped(json, false);
}

export function GetLatestTemplateResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetLatestTemplateResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'template': json['template'],
    };
}

export function GetLatestTemplateResponseToJSON(value?: GetLatestTemplateResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'template': value.template,
    };
}