// frontend/web/src/services/API/proserve-wb-publishing-api/models/ValidateProductVersionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Publishing API
 * API in charge of publishing workbench products
 *
 * The version of the OpenAPI document: 2023-01-26T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ValidateProductVersionRequest
 */
export interface ValidateProductVersionRequest {
    /**
     * The template definition of the new product version
     * @type {string}
     * @memberof ValidateProductVersionRequest
     */
    versionTemplateDefinition: string;
}

/**
 * Check if a given object implements the ValidateProductVersionRequest interface.
 */
export function instanceOfValidateProductVersionRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "versionTemplateDefinition" in value;

    return isInstance;
}

export function ValidateProductVersionRequestFromJSON(json: any): ValidateProductVersionRequest {
    return ValidateProductVersionRequestFromJSONTyped(json, false);
}

export function ValidateProductVersionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ValidateProductVersionRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'versionTemplateDefinition': json['versionTemplateDefinition'],
    };
}

export function ValidateProductVersionRequestToJSON(value?: ValidateProductVersionRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'versionTemplateDefinition': value.versionTemplateDefinition,
    };
}